# docs/openapi/rbac.yaml
openapi: 3.1.0
info:
  title: HandLab API - RBAC (Users, Organizations, Roles, Permissions, Memberships)
  description: |
    Contrato de API para o módulo RBAC (Role-Based Access Control) do sistema HandLab.
    
    ## Recursos
    - **Users**: Cadastro de usuários do sistema
    - **Organizations**: Organizações/clubes (V1: clube único - R34)
    - **Roles**: Catálogo de papéis (superadmin, dirigente, coordenador, treinador, atleta)
    - **Permissions**: Catálogo de permissões por papel
    - **Memberships**: Vínculos user↔organization+role
    
    ## Regras de Negócio
    - R25/R26: Permissões por papel e escopo
    - R29/R33: Exclusão lógica e histórico com rastro
    - R31/R32: Ações críticas auditadas
    - R34: Clube único na V1 (contexto organizacional obrigatório)
    - R42: Usuários sem vínculo ativo não podem operar
    
    ## Autenticação
    Todos os endpoints requerem autenticação via JWT Bearer token.
  version: 1.0.0
  contact:
    name: HandLab Team
  license:
    name: Proprietary
servers:
  - url: /v1
    description: API v1
security:
  - bearerAuth: []
...
paths:
...
  /organizations:
    get:
      operationId: listOrganizations
      tags: [organizations]
      summary: Listar organizações (paginado)
      description: |
        Lista paginada de organizações/clubes.
        
        **Nota V1:** Clube único (R34) — contexto organizacional implícito no token.
        
        **Regras:** R34 (clube único), R25/R26 (permissões)
      x-rule-ids: [R34, R25, R26]
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
        - name: order_by
          in: query
          schema:
            type: string
            enum: [created_at, name, updated_at]
            default: created_at
        - name: order_dir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: search
          in: query
          description: Busca por name ou code
          schema:
            type: string
      responses:
        '200':
          description: Lista de organizações
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationPaginatedResponse'
              example:
                items:
                  - id: "660e8400-e29b-41d4-a716-446655440001"
                    name: "Clube Handebol São Paulo"
                    code: "CHSP"
                    created_at: "2024-01-01T00:00:00Z"
                    updated_at: "2024-01-01T00:00:00Z"
                page: 1
                limit: 50
                total: 1
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

    post:
      operationId: createOrganization
      tags: [organizations]
      summary: Criar organização
      description: |
        Cria uma nova organização/clube.
        
        **Regras:** R25/R26 (permissões), R29 (exclusão lógica), R31/R32 (auditoria)
      x-rule-ids: [R25, R26, R29, R31, R32]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationCreate'
            example:
              name: "Clube Handebol São Paulo"
              code: "CHSP"
      responses:
        '201':
          description: Organização criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          description: Conflito de unicidade (code/name duplicado)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: conflict_unique
                message: "Código ou nome já cadastrado"
                details:
                  constraint: "organizations_code_key"
        '422':
          $ref: '#/components/responses/ValidationError'
  /organizations/{organization_id}:
    get:
      operationId: getOrganizationById
      tags: [organizations]
      summary: Obter organização por ID
      description: |
        Retorna os dados de uma organização específica.
        
        **Regras:** R25/R26 (permissões)
      x-rule-ids: [R25, R26]
      parameters:
        - $ref: '#/components/parameters/organizationIdParam'
      responses:
        '200':
          description: Dados da organização
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      operationId: updateOrganization
      tags: [organizations]
      summary: Atualizar organização
      description: |
        Atualiza dados de uma organização existente.
        
        **Regras:** R25/R26 (permissões), R31/R32 (auditoria)
      x-rule-ids: [R25, R26, R31, R32]
      parameters:
        - $ref: '#/components/parameters/organizationIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationUpdate'
            example:
              name: "Clube Handebol SP - Atualizado"
      responses:
        '200':
          description: Organização atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          description: Conflito de unicidade
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: conflict_unique
                message: "Código ou nome já cadastrado"
                details:
                  constraint: "organizations_code_key"
        '422':
          $ref: '#/components/responses/ValidationError'

  # ===========================================================================
  # ROLES (Catálogo - somente leitura)
  # ===========================================================================
...
  /organizations/{organization_id}/memberships:
    get:
      operationId: listMembershipsByOrganization
      tags: [memberships]
      summary: Listar vínculos por organização (paginado)
      description: |
        Lista paginada de vínculos (memberships) de uma organização.
        
        **Regras:** R25/R26 (permissões), R42 (vínculo ativo), R34 (clube único)
      x-rule-ids: [R25, R26, R42, R34]
      parameters:
        - $ref: '#/components/parameters/organizationIdParam'
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
        - name: is_active
          in: query
          description: Filtrar por status ativo/inativo
          schema:
            type: boolean
        - name: role_code
          in: query
          description: Filtrar por papel
          schema:
            type: string
            enum: [superadmin, dirigente, coordenador, treinador, atleta]
        - name: order_by
          in: query
          schema:
            type: string
            enum: [created_at, updated_at]
            default: created_at
        - name: order_dir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Lista de vínculos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipPaginatedResponse'
              example:
                items:
                  - id: "770e8400-e29b-41d4-a716-446655440002"
                    user_id: "550e8400-e29b-41d4-a716-446655440000"
                    organization_id: "660e8400-e29b-41d4-a716-446655440001"
                    role_code: "treinador"
                    is_active: true
                    created_at: "2024-01-15T10:30:00Z"
                    updated_at: "2024-01-15T10:30:00Z"
                page: 1
                limit: 50
                total: 1
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

    post:
      operationId: createMembershipForOrganization
      tags: [memberships]
      summary: Criar vínculo para organização
      description: |
        Cria um novo vínculo (membership) user↔organization+role.
        
        **Constraint:** UNIQUE(user_id, organization_id) onde is_active=true
        
        **Regras:** R25/R26 (permissões), R29 (exclusão lógica), R31/R32 (auditoria)
      x-rule-ids: [R25, R26, R29, R31, R32]
      parameters:
        - $ref: '#/components/parameters/organizationIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MembershipCreate'
            example:
              user_id: "550e8400-e29b-41d4-a716-446655440000"
              role_code: "treinador"
              is_active: true
      responses:
        '201':
          description: Vínculo criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Membership'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          description: Conflito de unicidade (vínculo ativo duplicado)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: conflict_unique
                message: "Usuário já possui vínculo ativo nesta organização"
                details:
                  constraint: "memberships_user_org_active_key"
        '422':
          $ref: '#/components/responses/ValidationError'
  /memberships/{membership_id}:
    get:
      operationId: getMembershipById
      tags: [memberships]
      summary: Obter vínculo por ID
      description: |
        Retorna os dados de um vínculo específico.
        
        **Regras:** R25/R26 (permissões)
      x-rule-ids: [R25, R26]
      parameters:
        - $ref: '#/components/parameters/membershipIdParam'
      responses:
        '200':
          description: Dados do vínculo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Membership'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      operationId: updateMembership
      tags: [memberships]
      summary: Atualizar vínculo
      description: |
        Atualiza role_code e/ou is_active de um vínculo existente.
        
        **Nota:** user_id e organization_id não podem ser alterados via API.
        
        **Regras:** R25/R26 (permissões), R29 (exclusão lógica), R31/R32 (auditoria)
      x-rule-ids: [R25, R26, R29, R31, R32]
      parameters:
        - $ref: '#/components/parameters/membershipIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MembershipUpdate'
            example:
              role_code: "coordenador"
              is_active: true
      responses:
        '200':
          description: Vínculo atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Membership'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          description: Conflito de unicidade (ativação duplicada)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: conflict_unique
                message: "Usuário já possui vínculo ativo nesta organização"
                details:
                  constraint: "memberships_user_org_active_key"
        '422':
          $ref: '#/components/responses/ValidationError'
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido via autenticação

  # ===========================================================================
  # Parameters
  # ===========================================================================
  parameters:
    pageParam:
      name: page
      in: query
      description: Número da página (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
    limitParam:
      name: limit
      in: query
      description: Itens por página (máx 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
    organizationIdParam:
      name: organization_id
      in: path
      required: true
      description: ID único da organização
      schema:
        type: string
        format: uuid
    membershipIdParam:
      name: membership_id
      in: path
      required: true
      description: ID único do vínculo
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Organization:
      type: object
      description: Organização/clube
      required:
        - id
        - name
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
          description: ID único da organização
        name:
          type: string
          minLength: 2
          maxLength: 255
          description: Nome da organização
        code:
          type: string
          nullable: true
          maxLength: 50
          description: Código curto (sigla)
        deleted_at:
          type: string
          format: date-time
          nullable: true
          readOnly: true
          description: Data/hora de exclusão lógica (RDB4)
        deleted_reason:
          type: string
          nullable: true
          readOnly: true
          description: Motivo da exclusão lógica
        created_at:
          type: string
          format: date-time
          readOnly: true
          description: Data/hora de criação
        updated_at:
          type: string
          format: date-time
          readOnly: true
          description: Data/hora da última atualização
    OrganizationCreate:
      type: object
      description: Payload para criação de organização
      required:
        - name
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 255
          description: Nome da organização
        code:
          type: string
          nullable: true
          maxLength: 50
          description: Código curto (sigla)
    OrganizationUpdate:
      type: object
      description: Payload para atualização de organização
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 255
          description: Nome da organização
        code:
          type: string
          nullable: true
          maxLength: 50
          description: Código curto (sigla)
    OrganizationPaginatedResponse:
      type: object
      description: Resposta paginada de organizações
      required:
        - items
        - page
        - limit
        - total
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Organization'
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0

    # -------------------------------------------------------------------------
    # Role Schema
    # -------------------------------------------------------------------------
...
    Membership:
      type: object
      description: Vínculo user↔organization+role
      required:
        - id
        - user_id
        - organization_id
        - role_code
        - is_active
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
          description: ID único do vínculo
        user_id:
          type: string
          format: uuid
          description: ID do usuário vinculado
        organization_id:
          type: string
          format: uuid
          description: ID da organização
        role_code:
          type: string
          enum: [superadmin, dirigente, coordenador, treinador, atleta]
          description: Código do papel atribuído
        is_active:
          type: boolean
          description: Se o vínculo está ativo
        created_at:
          type: string
          format: date-time
          readOnly: true
          description: Data/hora de criação
        updated_at:
          type: string
          format: date-time
          readOnly: true
          description: Data/hora da última atualização
    MembershipCreate:
      type: object
      description: Payload para criação de vínculo
      required:
        - user_id
        - role_code
      properties:
        user_id:
          type: string
          format: uuid
          description: ID do usuário a vincular
        role_code:
          type: string
          enum: [dirigente, coordenador, treinador, atleta]
          description: Código do papel (superadmin não permitido via API)
        is_active:
          type: boolean
          default: true
          description: Se o vínculo inicia ativo
    MembershipUpdate:
      type: object
      description: Payload para atualização de vínculo
      properties:
        role_code:
          type: string
          enum: [dirigente, coordenador, treinador, atleta]
          description: Código do papel (superadmin não permitido via API)
        is_active:
          type: boolean
          description: Se o vínculo está ativo
    MembershipPaginatedResponse:
      type: object
      description: Resposta paginada de vínculos
      required:
        - items
        - page
        - limit
        - total
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Membership'
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0

    # -------------------------------------------------------------------------
    # Error Response
    # -------------------------------------------------------------------------
    ErrorResponse:
      type: object
      description: Resposta de erro padronizada
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Código do erro
        message:
          type: string
          description: Mensagem de erro
        details:
          type: object
          nullable: true
          description: Detalhes adicionais (regra, constraint, etc.)

  # ===========================================================================
  # Responses
  # ===========================================================================
  responses:
    UnauthorizedError:
      description: Token inválido ou ausente
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: unauthorized
            message: "Token inválido ou ausente"
            details: null
    ForbiddenError:
      description: Permissão insuficiente (R25/R26)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: permission_denied
            message: "Permissão insuficiente"
            details:
              regra: "R25"
    NotFoundError:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: not_found
            message: "Recurso não encontrado"
            details: null
    ValidationError:
      description: Erro de validação (payload inválido)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: validation_error
            message: "Erro de validação"
            details:
              errors:
                - field: "email"
                  message: "Email inválido"